requirements.txt
attrs==25.3.0
BeatPrints==1.1.3
black==24.10.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
flasgger==0.9.7.1
Flask==3.0.0
Flask-Caching==2.3.1
Flask-Limiter==3.10.1
fonttools==4.56.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
limits==4.0.1
lrclibapi==0.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
mypy-extensions==1.0.0
numpy==1.26.4
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
Pygments==2.19.1
pylette==4.0.1
python-dotenv==1.0.0
PyYAML==6.0.2
questionary==2.1.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.5.0
toml==0.10.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.17.2
